+++ src/test/resources/application-test.properties
@@ -20,6 +20,8 @@ logging.level.org.hibernate.type.descriptor.sql=TRACE
 
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
+spring.mail.username=${MAIL_USERNAME:brad@sandbox5e63039302174a10a277a5c046516b36.mailgun.org}
 management.health.mail.enabled=false
 
 app.verification-url=https://localhost:8443/api/auth/verify-email?token=
+app.password-reset-url=http://localhost:3000/reset-password
